+++ contracts/interfaces/IERC20MintableBurnable.sol
@@ -4,6 +4,7 @@ pragma solidity ^0.8.13;
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 interface IERC20MintableBurnable is IERC20 {
-    function mint(address to, uint amount) external;
-    function burn(uint amount) external;
+    function mint(address to, uint256 amount) external;
+
+    function burn(uint256 amount) external;
 }
